Split test to external crate
