+++ trie-db/src/fatdb.rs
@@ -1,4 +1,4 @@
-// Copyright 2017, 2018 Parity Technologies
+// Copyright 2017, 2020 Parity Technologies
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -123,29 +123,5 @@ where
 					)
 				})
 			})
-	}
-}
-
-#[cfg(test)]
-mod test {
-	use memory_db::{MemoryDB, HashKey};
-	use crate::DBValue;
-	use keccak_hasher::KeccakHasher;
-	use reference_trie::{RefFatDBMut, RefFatDB, Trie, TrieMut};
-
-	#[test]
-	fn fatdb_to_trie() {
-		let mut memdb = MemoryDB::<KeccakHasher, HashKey<_>, DBValue>::default();
-		let mut root = Default::default();
-		{
-			let mut t = RefFatDBMut::new(&mut memdb, &mut root);
-			t.insert(&[0x01u8, 0x23], &[0x01u8, 0x23]).unwrap();
-		}
-		let t = RefFatDB::new(&memdb, &root).unwrap();
-		assert_eq!(t.get(&[0x01u8, 0x23]).unwrap().unwrap(), vec![0x01u8, 0x23]);
-		assert_eq!(
-			t.iter().unwrap().map(Result::unwrap).collect::<Vec<_>>(),
-			vec![(vec![0x01u8, 0x23], vec![0x01u8, 0x23])]
-		);
 	}
 }
