+++ trie-db/src/sectriedb.rs
@@ -1,4 +1,4 @@
-// Copyright 2017, 2018 Parity Technologies
+// Copyright 2017, 2020 Parity Technologies
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -80,26 +80,5 @@ where
 		CError<L>
 	> {
 		TrieDB::iter(&self.raw)
-	}
-}
-
-#[cfg(test)]
-mod test {
-	use memory_db::{MemoryDB, HashKey};
-	use hash_db::Hasher;
-	use keccak_hasher::KeccakHasher;
-	use reference_trie::{RefTrieDBMut, RefSecTrieDB, Trie, TrieMut};
-	use crate::DBValue;
-
-	#[test]
-	fn trie_to_sectrie() {
-		let mut db = MemoryDB::<KeccakHasher, HashKey<_>, DBValue>::default();
-		let mut root = Default::default();
-		{
-			let mut t = RefTrieDBMut::new(&mut db, &mut root);
-			t.insert(&KeccakHasher::hash(&[0x01u8, 0x23]), &[0x01u8, 0x23]).unwrap();
-		}
-		let t = RefSecTrieDB::new(&db, &root).unwrap();
-		assert_eq!(t.get(&[0x01u8, 0x23]).unwrap().unwrap(), vec![0x01u8, 0x23]);
 	}
 }
