+++ trie-db/test/src/lib.rs
@@ -1,0 +1,39 @@
+// Copyright 2020 Parity Technologies
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+
+//! Tests for trie-db crate.
+
+#[cfg(test)]
+mod proof;
+#[cfg(test)]
+mod iterator;
+#[cfg(test)]
+mod triedb;
+#[cfg(test)]
+mod triedbmut;
+#[cfg(test)]
+mod sectriedb;
+#[cfg(test)]
+mod sectriedbmut;
+#[cfg(test)]
+mod fatdb;
+#[cfg(test)]
+mod fatdbmut;
+#[cfg(test)]
+mod recorder;
+#[cfg(test)]
+mod iter_build;
+#[cfg(test)]
+mod trie_codec;
