+++ trie-root/test/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "trie-root-test"
+version = "0.16.0"
+authors = ["Parity Technologies <admin@parity.io>"]
+description = "Tests fo trie-root crate"
+repository = "https://github.com/paritytech/trie"
+license = "Apache-2.0"
+categories = [ ]
+edition = "2018"
+
+[dependencies]
+trie-root = { path = "..", version = "0.16.0" }
+hash-db = { path = "../../hash-db", version = "0.15.2" }
+hex-literal = "0.3"
+keccak-hasher = { path = "../../test-support/keccak-hasher", version = "0.15.2" }
+trie-standardmap = { path = "../../test-support/trie-standardmap", version = "0.15.2" }
+reference-trie = { path = "../../test-support/reference-trie", version = "0.22.0" }
